+++ src/main/java/com/example/cook/global/oauth2/OAuthAttributes.java
@@ -1,0 +1,92 @@
+package com.example.cook.global.oauth2;
+
+
+import com.example.cook.user.Role;
+import com.example.cook.user.SocialType;
+import com.example.cook.user.User;
+import com.example.cook.global.oauth2.userinfo.GoogleOAuth2UserInfo;
+import com.example.cook.global.oauth2.userinfo.KakaoOAuth2UserInfo;
+import com.example.cook.global.oauth2.userinfo.NaverOAuth2UserInfo;
+import com.example.cook.global.oauth2.userinfo.OAuth2UserInfo;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.Map;
+import java.util.UUID;
+
+/**
+ * 각 소셜에서 받아오는 데이터가 다르므로 소셜별로 데이터를 받는 데이터를 분기 처리하는 DTO 클래스
+ */
+@Getter
+public class OAuthAttributes {
+
+  private String nameAttributeKey; // OAuth2 로그인 진행 시 키가 되는 필드 값, PK와 같은 의미
+  private OAuth2UserInfo oauth2UserInfo; // 소셜 타입별 로그인 유저 정보(닉네임, 이메일, 프로필 사진 등등)
+
+  @Builder
+  public OAuthAttributes(String nameAttributeKey, OAuth2UserInfo oauth2UserInfo) {
+    this.nameAttributeKey = nameAttributeKey;
+    this.oauth2UserInfo = oauth2UserInfo;
+  }
+
+  /**
+   * SocialType에 맞는 메소드 호출하여 OAuthAttributes 객체 반환
+   * 파라미터 : userNameAttributeName -> OAuth2 로그인 시 키(PK)가 되는 값 / attributes : OAuth 서비스의 유저 정보들
+   * 소셜별 of 메소드(ofGoogle, ofKaKao, ofNaver)들은 각각 소셜 로그인 API에서 제공하는 회원의 식별값(id), attributes, nameAttributeKey를 저장 후 build
+   */
+  public static OAuthAttributes of(SocialType socialType,
+      String userNameAttributeName, Map<String, Object> attributes) {
+
+    /**
+     * 소셜 타입 별로 나눠서 빌더로 OAuthAttributes 빌드 시
+     * 유저 정보 추상 클래스인 OAuth2UserInfo 필드에 각 소셜 타입의 OAuth2UserInfo를 생성하여 빌드
+     */
+    if (socialType == SocialType.NAVER) {
+      return ofNaver(userNameAttributeName, attributes);
+    }
+    if (socialType == SocialType.KAKAO) {
+      return ofKakao(userNameAttributeName, attributes);
+    }
+    return ofGoogle(userNameAttributeName, attributes);
+  }
+
+  private static OAuthAttributes ofKakao(String userNameAttributeName, Map<String, Object> attributes) {
+    return OAuthAttributes.builder()
+        .nameAttributeKey(userNameAttributeName)
+        .oauth2UserInfo(new KakaoOAuth2UserInfo(attributes))
+        .build();
+  }
+
+  public static OAuthAttributes ofGoogle(String userNameAttributeName, Map<String, Object> attributes) {
+    return OAuthAttributes.builder()
+        .nameAttributeKey(userNameAttributeName)
+        .oauth2UserInfo(new GoogleOAuth2UserInfo(attributes))
+        .build();
+  }
+
+  public static OAuthAttributes ofNaver(String userNameAttributeName, Map<String, Object> attributes) {
+    return OAuthAttributes.builder()
+        .nameAttributeKey(userNameAttributeName)
+        .oauth2UserInfo(new NaverOAuth2UserInfo(attributes))
+        .build();
+  }
+
+  /**
+   * 이후에 CustomOAuth2UserService에서 DB에 저장할 내 서비스 User를 OAuth2UserInfo의 정보를 사용하여 빌더로 빌드 후 반환합니다.
+   * 앞서 CustomOAuth2User에서 언급한대로 email과 role도 사용하기 위해 추가하여 빌드합니다.
+   * email은 JWT Token을 발급하기 위한 용도뿐이므로 UUID를 사용하여 임의로 설정
+   *
+   * of메소드로 OAuthAttributes 객체가 생성되어, 유저 정보들이 담긴 OAuth2UserInfo가 소셜 타입별로 주입된 상태
+   * OAuth2UserInfo에서 socialId(식별값), nickname을 가져와서 build
+   * email에는 UUID로 중복 없는 랜덤 값 생성
+   * role은 GUEST로 설정
+   */
+  public User toEntity(SocialType socialType, OAuth2UserInfo oauth2UserInfo) {
+    return User.builder()
+        .socialType(socialType)
+        .socialId(oauth2UserInfo.getId())
+        .email(UUID.randomUUID() + "@socialUser.com")
+        .role(Role.GUEST)
+        .build();
+  }
+}
