+++ src/main/java/com/example/cook/global/oauth2/handler/OAuth2LoginFailureHandler.java
@@ -1,0 +1,26 @@
+package com.example.cook.global.oauth2.handler;
+
+/**
+ * OAuth2 로그인이 실패한다면, OAuth2LoginFailureHandler의 로직이 실행
+ */
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.stereotype.Component;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Slf4j
+@Component
+public class OAuth2LoginFailureHandler implements AuthenticationFailureHandler {
+  @Override
+  public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
+    response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+    response.getWriter().write("소셜 로그인 실패! 서버 로그를 확인해주세요.");
+    log.info("소셜 로그인에 실패했습니다. 에러 메시지 : {}", exception.getMessage());
+  }
+}
