+++ src/main/java/com/example/cook/user/service/UserService.java
@@ -6,6 +6,8 @@ import com.example.cook.user.dto.UserSignUpDto;
 import com.example.cook.user.repository.UserRepository;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -39,4 +41,5 @@ public class UserService {
     user.passwordEncode(passwordEncoder);
     userRepository.save(user);
   }
+
 }
