+++ src/main/resources/application-oauth.yml
@@ -22,7 +22,7 @@ spring:
             redirect-uri: http://localhost:8080/login/oauth2/code/kakao
             client-authentication-method: POST
             authorization-grant-type: authorization_code
-            scope: profile_nickname, profile_image
+            scope: profile_nickname
             client-name: Kakao
 
         provider:
