+++ src/main/resources/static/index.html
@@ -1,0 +1,3 @@
+<a href="/oauth2/authorization/kakao">Kakao Login</a><br>
+<a href="/oauth2/authorization/google">Google Login</a><br>
+<a href="/oauth2/authorization/naver">Naver Login</a><br>
